backend/requirements.txt
fastapi
uvicorn
langchain
chromadb
pypdf
tiktoken
python-dotenv
requests
httpx
sentence-transformers